feat: better contexts
