+++ apps/router/src/gateway-ui/types.ts
@@ -1,0 +1,3 @@
+export type GatewayConfig = {
+  baseUrl: string;
+};
