+++ apps/router/src/guardian-ui/hooks/index.tsx
@@ -1,33 +1,4 @@
-import { useContext, useEffect, useState } from 'react';
-import { GuardianAppContextValue, GuardianAppContext } from '../AppContext';
-import { SetupContext, SetupContextValue } from '../setup/SetupContext';
-import { AdminContext, AdminContextValue } from '../admin/AdminContext';
-
-export function useAppContext(): GuardianAppContextValue {
-  return useContext(GuardianAppContext);
-}
-
-export function useSetupContext(): SetupContextValue {
-  return useContext(SetupContext);
-}
-
-/**
- * Tells the guardian context to poll for updates. Handles turning off polling
- * on dismount.
- */
-export function useConsensusPolling(shouldPoll = true) {
-  const { toggleConsensusPolling } = useSetupContext();
-
-  useEffect(() => {
-    if (!shouldPoll) return;
-    toggleConsensusPolling(true);
-    return () => toggleConsensusPolling(false);
-  }, [shouldPoll, toggleConsensusPolling]);
-}
-
-export function useAdminContext(): AdminContextValue {
-  return useContext(AdminContext);
-}
+import { useEffect, useState } from 'react';
 
 export const useEllipsis = () => {
   const [ellipsis, setEllipsis] = useState('');
